ARSIS-Devices/VisionKit/control_endpoint/pyproject.toml
[tool.poetry]
name = "telemetry-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "telemetry_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
pyserial = "^3.5"
uvicorn = {extras = ["standard"], version = "^0.20.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
